Bootblock 0.5.2     https://github.com/Nuvoton-Israel/npcm8xx-bootblock/releases/tag/A35_BootBlock_0.5.2       * Bug fix: Errata fix: 1.7 eSPI FATAL_ERROR: Set ESPI_ESPI_ENG to 0x40 (remove RMW to set bit 6).       * Update CP init code (disabled by default, under SEARCH_CP_ON_FLASH_0 flag).       * In case of traps and unexpected IRQs and FIQ: clear GIC registers.       * In case of TRAP: notify TIP (TIP mode), or FSW (NO_TIP mode).       * Update GIC driver and GIC table.

+++ Sources/SWC_HAL/ModuleDrivers/shm/1/shm_drv.c
@@ -65,7 +65,7 @@ static SHM_ESPI_CALLBACK_T SHM_espiCallback;
 /*                                  LOCAL FUNCTIONS FORWARD DECLARATIONS                                   */
 /*---------------------------------------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------------------------------------*/
-void SHM_IntHandler     (UINT16 int_num);
+void SHM_IntHandler     (UINT int_num);
 
 /*---------------------------------------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------------------------------------*/
@@ -808,7 +808,7 @@ void SHM_ReleaseHostWait (void)
 /* Description:                                                                                            */
 /*                  This routine is the Shared Memory interrupt handler.                                   */
 /*---------------------------------------------------------------------------------------------------------*/
-void SHM_IntHandler (UINT16 int_num)
+void SHM_IntHandler (UINT int_num)
 {
     UINT8 smc_sts;
     UINT8 hofs_sts  = 0;
